Added the possibility to edit notes.
Updated the edit note button to open a new activity that gets the notes info via intent and sets it to be edited on the view. Then when the user presses save the note gets updated using sqlite and it's saved and the user is redirected back to the cardview activity.