Move voice recording and playing to the right places.
Also make so that when user presses the back button it leads the user to the MainActivity.
The voice recording is now made in the new note activity and it can be listened to in the notes view, and when the note is deleted, the file will also be deleted from the storage memory.