Add a tag possibility for notes.
Add a possibility to add custom tags to notes and create new tags. User can select an existing tag or create a new one and the selected tags are shown in a textfield below the button. They are not yet saved to a database. New tags are saved to sharedprefs for convenience.