User can now edit the tags on each note.
Each tag is displayed in a textview as a String list. When user presses the "edit tags" button the tags list is opened with the chosen tags selected already for the user. The user can update the selected tags and press ok, and when the edits are saved the tags are updated along with everything else.